BIOIN-203 modified function create_control_signature so that it now handles a vcf file rather than a bed file:
Creates a control signature that matches each SNP in the input signature vcf file with an adjacent position with
the same trinucleotide motif, maintaining the same ref and alt composition. Non-SNP entries are ignored.
Adds an ORIG_SEQ info in the output vcf indicating the position of the original variant this controls for.