clean calculation of residual snv rate
This commit cleans the way that the effective_bases_covered and
residual_snv_rate_no_filter are calculated in the srsnv report / plotting utils. Before,
residual_snv_rate_no_filter was calculated as the ratio of  the FP train/test dataset size
in the numerator of and effective_bases_covered as the denominator. The latter
included a factor of the downsampling rate (from the number of entries in the FP
featuremap file to the [smaller] dataset size). After this commit, the numerator is
the number of reads in the fp featuremap directly, and the downsampling rate is
removed from the denominator.

This commit also fixes a factor of 2 in the snv_rate that was caused by calculating the
downsampling rate as the ratio : (TP + FP dataset set) / (FP entry number).