Restrict training to chromosomes in TP bed file
This is needed for splitting the data into folds for cross-validation:
fold-splitting is done so that folds will all have approximately the
same size. If training is done only on a few chromosomes, should make
sure that they are not in the same fold, and that the fold sizes
are approximately equal.