[BIOIN-1448] Implemented filtering of VCF that changes GQ/GT/PL (#160)
This PR updates the way we are filtering GATK VCFs. Until now we trained the filtering models using approximate ground truths. This allows marking the variant as True/False but not changing the genotype.

In this PR I am adding the ability to use a model that is trained on the exact ground truth and in this case it is possible to not only filter the variant but also change the genotype. The model works (similar to the DeepVariant) by predicting genotype one of the three possible genotypes: hom-ref, hom-alt and het-alt.

The main challenge that also required some code work was applying the model to multiallelic variants. The heuristic that I chose for the multiallelic variants: R,A1,A2 (the first one is the reference).

Pick the "strongest" allele (e.g. A1),
Combine the support for A1 and A2 into a single allele A
Genotype (R,A): the result is (PL'(R,R), PL'(R,A), PL'(A,A))
Genotype (A1,A2), where A1 is the reference, A2 is the "alt".
Merge the support in the following way to generate PL(R,R),PL(R,A1),PL(A1,A1),PL(R,A2),PL(A1,A2),PL(A2,A2) tuple
PL(R,R) = PL'(R,R)
PL(R,A1) = PL'(R,A)
PL(A1,A1) = PL'(A,A) * PL'(A1,A1)
PL(A1,A2) = PL'(A,A) * PL(A1,A2)
PL(A2,A2) = PL(A,A) * PL(A2,A2)

Note that PL(R,A2) is zero here. 

This logic is implemented in the function `variant_filtering_utils.combine_multiallelic_spandel`

There are two main files to review, everything else is not really important (cosmetics)
`filter_variants_pipeline_main.py`
`variant_filtering_utils.py`