SRSNV report fix auc bug and code cleanup
ROC AUC can only be calculated when the dataset is not empty,
and when both classes (True, False) are represented in the dataset.
This commit adds checks for these conditions before calculation
or ROC AUC. It also adds check whether any reads are mixed before
calculating mixed-only statistics.

In addition, the code for SRSNV report is cleaned from unnecessary
code belonging to old report.